+++ Proyecto1/Expresiones/literal.hpp
@@ -11,8 +11,8 @@ public:
     Type type;
     std::string strVal;
     bool boolVal;
-    float floatVal;
-    Literal(int, int, Type, std::string, int, bool, float);
+    double floatVal;
+    Literal(int, int, Type, std::string, int, bool, double);
     Symbol ejecutar(Environment *env) override;
 };
