+++ Proyecto1/Instrucciones/print.cpp
@@ -21,7 +21,7 @@ void Print::ejecutar(Environment *env) {
             break;
         case FLOAT:
             //tree->ConsoleOut += std::to_string(*static_cast<int*>(sym.Value))+ "\n";
-            single->addConsola(std::to_string(*static_cast<float*>(symb.value)) + "\n");
+            single->addConsola(std::to_string(*static_cast<double*>(symb.value)) + "\n");
             break;
         case BOOL:
             if(*static_cast<bool*>(symb.value)){
